refactor: remove dto package in domain layer (#37)
