+++ api/src/main/kotlin/team/backend/quiz/presentation/QuizController.kt
@@ -4,7 +4,7 @@ import jakarta.validation.Valid
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
-import team.backend.quiz.domain.dto.command.QuizCommand
+import team.backend.quiz.domain.command.QuizCommand
 import team.backend.quiz.domain.service.QuizService
 import team.backend.quiz.presentation.dto.QuizDto
